pyproject.toml
[tool.poetry]
name = "intellidoc"
version = "0.1.0"
description = ""
authors = ["Moitree Basu <sfurti.basu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
pandas = "^2.2.3"
pypdf2 = "^3.0.1"
tqdm = "^4.66.6"
torchtext = "^0.18.0"
nltk = "^3.9.1"
scikit-learn = "^1.5.2"
transformers = "^4.46.1"
pytest = "^8.3.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
mypy = "^1.13.0"
pytest-cov = "^6.0.0"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
